Refactoring files to be more idiomatic
