--- src/pokemon-service/usecase/presenter/pokemon_presenter.go
@@ -1,11 +1,0 @@
-package presenter
-
-import (
-	"github.com/GerardoHP/ondemand-go-bootcamp/domain/entity"
-)
-
-// Specific  implementation of output for use cases
-type PokemonPresenter interface {
-	ResponsePresenter(p []*entity.Pokemon) []*entity.Pokemon
-	IndividualResponsePresenter(p *entity.Pokemon) *entity.Pokemon
-}
